+++ src/Regressors/KNNRegressor.php
@@ -173,7 +173,7 @@ class KNNRegressor implements Online, Persistable
         $predictions = [];
 
         foreach ($dataset as $sample) {
-            [$distances, $labels] = $this->neighbors($sample);
+            [$distances, $labels] = $this->nearest($sample);
 
             if ($this->weighted) {
                 $weights = [];
@@ -200,7 +200,7 @@ class KNNRegressor implements Online, Persistable
      * @param array $sample
      * @return array[]
      */
-    protected function neighbors(array $sample) : array
+    protected function nearest(array $sample) : array
     {
         $distances = [];