+++ tests/Clusterers/KMeansTest.php
@@ -20,9 +20,9 @@ use RuntimeException;
 
 class KMeansTest extends TestCase
 {
-    protected const TRAIN_SIZE = 400;
+    protected const TRAIN_SIZE = 350;
     protected const TEST_SIZE = 10;
-    protected const MIN_SCORE = 0.9;
+    protected const MIN_SCORE = 0.8;
 
     protected $generator;