+++ tests/Regressors/ExtraTreeRegressorTest.php
@@ -31,7 +31,7 @@ class ExtraTreeRegressorTest extends TestCase
     {
         $this->training = unserialize(file_get_contents(dirname(__DIR__) . '/mpg.dataset') ?: '');
 
-        $this->testing = $this->training->randomize()->head(self::TEST_SIZE);
+        $this->testing = $this->training->randomize()->take(self::TEST_SIZE);
 
         $this->estimator = new ExtraTreeRegressor(30, 3, 0., null, 1e-4);