+++ tests/Regressors/RegressionTreeTest.php
@@ -17,7 +17,7 @@ use RuntimeException;
 
 class RegressionTreeTest extends TestCase
 {
-    protected const TRAIN_SIZE = 300;
+    protected const TRAIN_SIZE = 400;
     protected const TEST_SIZE = 10;
     protected const MIN_SCORE = 0.7;